+++ app/src/main/java/com/danielpriddle/drpnews/ui/ArticleListComposable.kt
@@ -1,4 +1,4 @@
-package com.danielpriddle.drpnews.ui.composables
+package com.danielpriddle.drpnews.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
